notes now uses musicxml
no jank stuff anymore
