--- src/Melodic Analysis/midi.py
@@ -1,9 +1,0 @@
-    # f.close()
-
-    # xmldoc = minidom.parse(f)
-    # print(xmldoc[0])
-    # for element in f.recurse():
-    #     msg = str(element)
-    #     print(msg)
-    # f.close()
-    # notes = dom.getElementsByTagName("note")
